Add files via upload
This is a dataset on mobility trends during Covid-19. The data was collected through a period of 3 months (Jan-March) in several countries and cities around the world by Apple. This dataset was made available by data.world. Data wrangling was performed on the dataset in R to make the data ready for analysis.